Implement sand grains and experience updates at attendance

+++ src/main/java/com/junsu/cyr/model/user/UserSidebarResponse.java
@@ -8,7 +8,7 @@ import lombok.Data;
 public class UserSidebarResponse {
     private Integer userId;
     private Long expCnt;
-    private Long sand;
-    private Long glass;
+    private Integer sand;
+    private Integer glass;
     private Integer temperature;
 }
